requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
tavily-python>=0.3.3
python-dotenv>=1.0.0
